pyproject.toml
[project]
name = "voice-caption"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20250625",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "psutil>=7.0.0",
]
